Moved from single mutex to recursive.
Why?
Inverters need to take modbus_device lock to make sure the ModbusNode task does not update the device.
Having the lock before the driver executes gives us time to make all we want to do on that loop inside the driver. Also if the inverter is not able to get the lock, it will just retry next time, when the modbusNode releases the device.

Inside the driver implementation we have situations where we add requests/delete requests from the modbus. these are locked/unlocked on function call of the device to make sure they are thread safe.